+++ OMICALL-Flutter-SDK-Sample/lib/screens/home/home_screen.dart
@@ -21,8 +21,10 @@ class HomeScreen extends StatefulWidget {
 }
 
 class _HomeScreenState extends State<HomeScreen> {
+  // late final TextEditingController _phoneNumberController =
+  //     TextEditingController()..text = Platform.isAndroid ? '111' : '112';
   late final TextEditingController _phoneNumberController =
-      TextEditingController()..text = Platform.isAndroid ? '111' : '112';
+  TextEditingController()..text = Platform.isAndroid ? '123aaa' : '122aaa';
   TextStyle basicStyle = const TextStyle(
     color: Colors.white,
     fontSize: 16,
@@ -44,11 +46,9 @@ class _HomeScreenState extends State<HomeScreen> {
   @override
   void initState() {
     super.initState();
-    if (widget.needRequestNotification) {
-      updateToken();
-    }
-    _subscription = OmicallClient.instance.controller.eventTransferStream
-        .listen((omiAction) {
+    updateToken();
+    _subscription =
+        OmicallClient.instance.callStateChangeEvent.listen((omiAction) {
       if (omiAction.actionName == OmiEventList.incomingReceived) {
         //having a incoming call
         final data = omiAction.data;
@@ -270,7 +270,11 @@ class _HomeScreenState extends State<HomeScreen> {
         status: CallStatus.calling,
       );
     }
-    OmicallClient.instance.startCall(
+    // OmicallClient.instance.startCall(
+    //   phone,
+    //   _isVideoCall,
+    // );
+    OmicallClient.instance.startCallWithUUID(
       phone,
       _isVideoCall,
     );