+++ android/src/main/kotlin/vn/vihat/omicall/omicallsdk/constants/constanst.kt
@@ -1,7 +1,10 @@
 package vn.vihat.omicall.omicallsdk.constants
 
 //EVENT
-const val INIT_CALL = "INIT_CALL"
+const val START_SERVICES = "START_SERVICES"
+const val CONFIG_NOTIFICATION = "CONFIG_NOTIFICATION"
+const val INIT_CALL_API_KEY = "INIT_CALL_API_KEY"
+const val INIT_CALL_USER_PASSWORD = "INIT_CALL_USER_PASSWORD"
 const val UPDATE_TOKEN = "UPDATE_TOKEN"
 const val START_CALL = "START_CALL"
 const val END_CALL = "END_CALL"
@@ -17,12 +20,15 @@ const val INPUTS = "INPUTS"
 const val OUTPUTS = "OUTOUTS"
 const val SET_INPUT = "SET_INPUT"
 const val SET_OUTPUT = "SET_OUTPUT"
+const val START_CALL_WITH_UUID = "START_CALL_WITH_UUID"
+const val LOG_OUT = "LOG_OUT"
 
 //LISTENER
 const val CALL_ESTABLISHED = "CALL_ESTABLISHED"
 const val CALL_END = "CALL_END"
 const val INCOMING_RECEIVED = "INCOMING_RECEIVED"
 //const val CONNECTION_TIMEOUT = "CONNECTION_TIMEOUT"
-const val HOLD = "HOLD"
+const val SPEAKER = "SPEAKER"
+const val VIDEO = "VIDEO"
 const val MUTED = "MUTED"