+++ lib/constant/names.dart
@@ -2,7 +2,10 @@
 // ignore_for_file: constant_identifier_names
 
 class OmiActionName {
-  static const String INIT_CALL = "INIT_CALL";
+  static const String START_SERVICES = "START_SERVICES";
+  static const String INIT_CALL_API_KEY = "INIT_CALL_API_KEY";
+  static const String CONFIG_NOTIFICATION = "CONFIG_NOTIFICATION";
+  static const String INIT_CALL_USER_PASSWORD = "INIT_CALL_USER_PASSWORD";
   static const String UPDATE_TOKEN = "UPDATE_TOKEN";
   static const String START_CALL = "START_CALL";
   static const String END_CALL = "END_CALL";
@@ -18,4 +21,6 @@ class OmiActionName {
   static const String INPUTS = "INPUTS";
   static const String SET_INPUT = "SET_INPUT";
   static const String SET_OUTPUT = "SET_OUTPUT";
+  static const String START_CALL_WITH_UUID = "START_CALL_WITH_UUID";
+  static const String LOG_OUT = "LOG_OUT";
 }