+++ CHANGELOG.md
@@ -2,25 +2,42 @@
 
 All notable changes to this project will be documented in this file.
 
+##2.2.1
+- **BREAKING CHANGE**
+  - We support call with uuid `Omiclient.instance.startCallWithUUID`
+  - Increase Android core version (fix crash on some devices)
+  - Update sample
 
+##2.1.0
+- **BREAKING CHANGE**
+  - We support config notification `Omiclient.instance.configPushNotification`, But it doesn't support on iOS
+  - Increase Android core version
+  - Update sample
 
-## [1.0.0](https://github.com/VIHATTeam/OMICALL-Flutter-SDK.git) (13/12/2022)
+##2.0.2
+- Fix crash on release mode: login and update notification for Android
+- Increase core version
+- Add `startServices` function
 
+##2.0.1
+- Support API for SDK
+- Increase core version
+- Update documents
 
 ## 2.0.0
-* Support App - to - App
-* Support Video Call
-* Improve performance and logic
+- Support App - to - App
+- Support Video Call
+- Improve performance and logic
 
 ## 1.0.12
 
-* Support iOS 11,12
-* Support minSDK from 21.
+- Support iOS 11,12
+- Support minSDK from 21.
 
 ## 1.0.11
 
-* Add more event list
-* Refactor code
+- Add more event list
+- Refactor code
 
 
 ## 1.0.10